spinlock contention made long

+++ src/include/kernel/spinlock.h
@@ -32,7 +32,7 @@
 #define SPIN_CONTENTION_LIMIT 0x10000000
 
 #ifdef SPIN_CONTENTION_LIMIT
-# define __SPIN_CONTENTION_FIELD      unsigned int contention_count;
+# define __SPIN_CONTENTION_FIELD      unsigned long contention_count;
 # define __SPIN_CONTENTION_FIELD_INIT contention_count : SPIN_CONTENTION_LIMIT,
 #else /* SPIN_CONTENTION_LIMIT */
 # define __SPIN_CONTENTION_FIELD
